requirements.txt
requests~=2.32.3
Flask~=3.1.0
Flask-Cors~=5.0.0
Flask-Caching~=2.3.0
python-dotenv~=1.0.1
streamlit~=1.41.1